5/31 基本就是毕设最后的结构了，有问题再修。

+++ Plugins/HITerrainMaker/Source/HITerrainMaker/Private/TerrainMaths/Modules/HITerrainPerlin.cpp
@@ -48,7 +48,7 @@ void FHITerrainPerlin::SetPersistence(float InPersistence)
 	Perlin.SetPersistence(Persistence);
 }
 
-void FHITerrainPerlin::SetOctaveCount(float InOctaveCount)
+void FHITerrainPerlin::SetOctaveCount(int32 InOctaveCount)
 {
 	OctaveCount = InOctaveCount;
 	Perlin.SetOctaveCount(OctaveCount);
